roles: refactor 'rpm_ostree_status*' to use json_query filter
This attempts to streamline the `rpm_ostree_status` role by employing
the `json_query` filter.  We can query the `deployments` from the
output of `rpm-ostree status --json` for the correct deployment which
is booted or not booted.  In the case of a single deployment, we
default to setting the `ros_not_booted` to false.

The `rpm_ostree_status_verify` has been converted to use the
json_query filter as well.  In this role, the `rpm_ostree_status` is
symlinked into the directory, so it can be included directly.  This
was done because we want to have that role run every time the 'verify'
role is called.  (I tried to use the meta dependencies, but it would
only run once per playbook if the 'verify' role was called multiple
times)

I also snuck it a little fix to the `admin-unlock` test that I noticed
while checking to see if the all the tests still worked.

Closes #230