Add atomic CLI tests
This commit adds tests for the the atomic command.  The tests revolve
mostly around pulling images and installing/running containers with
variations of the options in the atomic subcommand.  These tests include:

  - Pull by fully qualified image name
  - Pull by short image name
  - Pull by fully qualified image name with --name option
  - Pull with tags
  - Pull by digest
  - Pull to ostree/docker storage
  - Images update command

Most of the tests use the cockpit container.  This container was chosen
because it has most of the labels that the atomic command utilizes. For
simple tests, the busybox container was chosen because it is small.