roles: enable allow_duplicates for rpm_ostree_status* roles
This changes the `rpm_ostree_status*` roles to be used more than once
in a playbook via the `allow_duplicates` toggle.

The side-effect is that the tests that use those roles multiple times
have to use them as roles or via `include_role:`.  So this change also
modifies the existing tests to adapt to this new requirement.