tests: grab commits at beginnig of each 'rpm-ostree' test
We can run into trouble when running the `rpm-ostree` test if the
stream under test gets a new commit during the course of the test.  To
mitigate this, we pull the `HEAD` and `HEAD-1` commits before running
the bulk of each test.  This way the test will use those two commits
during the test, rather than getting an updated/new commit in the
middle of things.