tests: use the 'reboot' role
Update the two tests to use the `reboot` role rather than explcitly
including the file.

Additionally, change the `rpm-ostree` test to use the parent of the
deployed commit as a stand-in for the `HEAD-1` commit.  The `HEAD-1`
commit that is fetched from the remote can be different from the
parent of the deployed commit, so this avoids any changes to the
remote that might upset the test.