roles:  use new regexp method to append kube options
The previous approach to adding required options to the `KUBELET_ARGS`
did not take into account that there could be default options passed
in as args.  This changes uses the `lineinfile` module and its regexp
capabilities to match the `KUBELET_ARGS` line in the file and append
the options we want to the existing line.

Additionally, the `--config` flag for the kubelet was marked as
'Deprecated' way back in Aug 2016[0] and then completely removed in Jan
2017[1].  The replacement option is `--pod-manifest-path`.

[0] https://github.com/kubernetes/kubernetes/commit/e43ccdbf2c0b69bfa19108a4d82894962c108016
[1] https://github.com/kubernetes/kubernetes/commit/bec6635ccc05b1f02c9971ec9bf529faac0b9ca1