roles: print matching lines from journal
Previously, when the `journal_search` role would fail, it would only
print a message saying that the matching line had been found.  This is
only partially useful if you want to see the actual matching line in
the test output.

This changes the `journal_search` role to crudely print the matching
lines by way of running a `grep` against the output of `journalctl`
after a match has been found.  The reason for doing this operation
after the match has been found is because the command `journal | grep
<string>` may end up in the journal if it is run with elevated
privileges via Ansible (i.e. using `sudo`)

Closes #253