journal_search: improve detection of messages in journal
We can exclude some of the messages in the journal that are generated by
running Ansible tasks/playbooks by just piping the output into an
inverse `grep` search.  This improves the reliability of the messages
getting matched by the role.

Additionally, this introduces support for passing extra arguments to
the `journalctl` command.  This allows users to be even more specific
with the journal messages they want to match; for example, matching
AVC denials using the `_TRANSPORT=audit` field.