tests: handle package layering on CentOS
In #254, it is noted that the package layering of `httpd` isn't
working on CentOS because the `centos-logos` dependency is trying to
layer files on `/boot`.  Until CentOS gets a version of `rpm-ostree`
that supports layering on `/boot` (see projectatomic/rpm-ostree#969),
we need to use a different package to test package layering in CentOS.

This change brings in a conditional installation of `ntp` in the
CentOS case, until we have a newer version of `rpm-ostree` to use.