roles/tests: handle shared variable namespace
When running a role multiple times in a playbook, we can run into
problems with variables since they all live in the same namespace.
Notably, when you run a role multiple times, variables that are
optional but have been defined, will carry over to the next execution
of the role.  Or if two roles have the same variable name, you can get
values stomping on each other.

This commit tries to alleviate some of these problems by updating the
variable names in the roles to be unique to the role.  Additionally,
when using roles multiple times in playbooks, both required and optional
variables are explicitly defined to avoid any overwriting.