roles/bdsf: prepend 'ansible_host' to archive file
If you ran a long test that used the `booted_deployment_set_fact` role
across multiple hosts in parallel, any archive file that was created
by the test would be clobbered by the last execution that completed.
This creates a problem when you are trying to read that file and
expecting it to have certain values.

To avoid this, the archive file is created with the value of
`ansible_host` prepended to the file name.  This keeps the file unique
to the host under test and should avoid any clobbering that may have
happened.

Closes #258