system-containers: disable etcd/flanneld if rpms are installed
On certain flavors of Atomic Host, we still ship the `etcd` and
`flannel` RPMs on the host.  This causes a minor conflict when we
install those services from system containers.  Notably, the existing
services can be running and attempts to `runc exec` into the
containers will fail because they haven't been started yet.

So we check for the presence of the RPMs and proactively disable/stop
the services before trying to use the system containers.