selinux_verify: log any files/procs with incorrect SELinux labels
Previously, the `selinux_verify` role would fail with a generic
message about finding a file or process that had an incorrect SELinux
label.  This caused additional work for anyone doing triage of the
tests and made folks unhappy.

This change alters how the role reports any instances of incorrect
SELinux labels.  The files/directories/processes are saved into a list
which can then be read out later as part of a failure message.

This should result in last failure message to include the relevant
details of the files/processes and negate the need to dig through the
logs further during triage.

Closes #260