Add test for livefs
This commit adds a simple livefs test using package layering as the
change in the filesystem.

The rpm_ostree_install_verify and rpm_ostree_uninstall_verify roles
needed an option to skip the rpm-ostree status since the output is
changed with the livefs implementation. The roles were also cleaned
up and now uses the new variables from the rpm_ostree_status role.