admin-unlock: use different RPM on Fedora
In #271, the RPM used during the sanity test of `ostree admin unlock`
was changed for the Fedora case.  We make the same changes here in the
more comprehesive test of the `ostree admin unlock` functionality.

As part of this, the base EPEL URL is updated (who knew it chagned?!)
and the `sed` regex was updated to be a little cleaner.