rpm_install: use dummy RPM to support cross-platform installs
Previously we had been able to use the `epel-release` RPM as a way to
test RPM installs when a host had used `ostree admin unlock`.  But the
`epel-release` RPM was changed to conflict with the `fedora-release`
RPM, meaning we were not able to use a single RPM across all
platforms.

Instead of trying to setup conditions for when to use `epel-release`
and when not to, we are now able to use a 'dummy' RPM that will
install across all platforms.  (Thanks @jlebon!)

This change implements the use of the 'dummy' RPM in the tests that
use the `rpm_install` role.

Additionally, I snuck in some minor changes to the `epel_rpm_url`
role:  updating the base EPEL URL and improving the `sed` regex for
readability.