rpm_install: implement installing RPM located on host
The `rpm_install` role currently operates by installing an RPM from a
URL provided.  With the addtion of the dummy RPM in #281, we can copy
the RPM to the host and perform the `rpm -i` using the file copied to
the host.

This change implements the ability to copy an RPM to the host and then
installing said RPM directly.  It retains the ability to install from
a URL, as that is still useful (and used) in other tests.