tests: handle 'atomic install --system' on old versions
Well, we're still using `fedora/26/atomic` in our CI checks and the
version of `atomic` there is friggin old.  How old?  So old that it
will try to land an RPM on the host when you do an 'atomic install
--system' by default.

This tries to work around this problem by doing a version check and
altering how the `atomic install` command is used depending on the
version detected.