Gather version information for common packages
The goal of atomic host tests is to run across the latest released
and development streams of Atomic Host (Fedora, RHEL, CentOS). Some
of these streams are slower to get newer packages and causes breakage
in the test due to missing/broken functionality.  We have dealing
with these by getting the version of the package as needed and
conditionally performing actions based on the version.

This PR aims to simplify this process by centralizing the gathering
of about the atomic host in one role and making the information
available through a dictionary.  Currently, there is only version
information about common packages, but anything can be added to the
dictionary.  It is pretty much like the setup module in Ansible but
specifically for Atomic Host.