Improve error messages in roles
Sometimes its not easy to figure out why a test failed based on
the error message.  Digging into the code to find the failure
comparison that caused the failure is required.

Lets make the error messages more detailed when comparisons are
made so triaging failures is easier.

Note: This commit is only improving the error messages in the roles.