add 'atomic_install_version_check' role
I was about to add a 3rd copy of the same code to detect which version
of `atomic` we are using when we run the tests.  But I came to my
senses and added a new role that handles all that.

The `atomic_install_version_check` role does a version check and sets
up a variable to use `atomic install`.  This is a follow-on (and
improvement, IMO) to #284.