docker-build-httpd:  set MTU equal to default interface
When running an OpenStack environment, the host under test might end
up with an MTU that doesn't jive with the Docker daemon (usually when
the MTU is lower than 1500).  This can lead to network issues when
doing `docker build`.  So we'll just set the MTU for the Docker
daemon to the same.