roles/rpm_ostree_status: exclude noisy keys
In the newer versions of `rpm-ostree`, the RPM DB package list is
included in the commit metadata.  This causes the output of
`rpm-ostree status --json` to be very noisy when run via Ansible (i.e.
100s of package key/values are now in the output).

This attempts to work around the noise by running the output through
`jq` and deleting the offending keys.