docker_private_registry: workaround missing atomic-registries
As noted in CentOS/sig-atomic-buildscripts#316, the downstream CentOS
Atomic Host and CAHC composes are missing the `atomic-registries`
package.  This is critical for the configuration of insecure
registries, as is necessary when setting up a registry in a container.

This works around the missing package by using package layering to
install the missing package, then proceeding with the remainder of the
role.

Closes #316