roles: convert docker roles to be more generic
We have a number of specialized roles that are used in the build, run,
and removal of an `httpd` container.  While this has worked thusfar,
it limits the flexibility around the build/run/remove operations of
Docker containers.

This change modifies the existing roles to be more generic and allows
us to use them with more flexibility.

As part of the changes, a new role named `check_open_port` was added
to replace some functionality that was removed from an exiting role.

Finally, the existing tests that used the roles were updated to use
the new role names.