i-s-t: add test for non-priv systemd container
Since it is possible to run systemd in a container without the
`--privileged` flag, we decided we should have a test for this going
forward.

This change adds in the files necessary to build an `httpd` container
that uses `systemd` as its init mechanism.  Additionally, we add the
building/running of said container to the `improved-sanity-test`.