selinux_verify: add tags for each label check
The `selinux_verify` role performs multiple checks of files and
processes on the host to make sure all the SELinux labels are
correct.  However, if we wanted to exclude a particular part of the
role (because of known issues), we would have to exclude the entire
role.

This change introduces a set of tags assigned to different tasks in
the role, so that we are able to selectively choose which checks
should run as part of the role.