k8s-cluster: fix test for F27AH
In Fedora 27 Atomic Host, all of the `kubernetes` related packages
were removed from the host.  The replacement functionality is now
provided by system containers.

This set of changes reworks the `kubernetes_setup` role to utilize the
new system containers to bring up a single node cluster and then
proceed with the remaininder of the `k8s-cluster` test.

NOTE:  the system containers are only used in the case where the
`kubernetes` packages are not present on the host.  For other
platforms where they are present on the host, the original method of
using manifests continues to be used.

Many thanks to Kedar Bidarkar for starting this work!