docker_rmi: rework to support tagged images
Previously, if a tagged image was passed to the role, it could fail
because the role would search for the string in the output of `docker
images`.  But the output of `docker images` uses a table format and
separates the image name from the tag.

This change reworks the role to specifically select the image name
from the output of `docker images` and then decides if output returned
is what is expected.  If we are looking for the image to be present,
`docker images` will return 2 lines of stdout.  Whereas, if we are
looking for the image to be missing, `docker images` will only return
1 line of stdout.

Additionally, the `docker rmi` command will now always use the `-f`
option to forcefully remove the image.