roles: new role to validate firewalld install/config
This introduces a new role which can be used to validate the
installation/configuration of the `firewalld` service.  It will
determine if the host under test should have `firewalld` installed and
how it should be configured if it is installed.

This initial attempt is very simple, but could be expanded to do more
complex checks if desired.

The role is used in the `improved-sanity-test` before the host is
upgraded and after upgrade.

Closes #315