ci: enable extra args to be passed to .test_director
The `centos/7/atomic` context has been consistently failing because
there are AVC denials from the kernel that are present in the version
of CentOS AH we are using.  This is preventing any useful CI checks
from being run on that platform.

This change introduces a simple check to the `.test_director` script to
see if any arguments were passed to the script.  This way, we are able
to skip tags we know are causing problems in our CI checks.

In order to utilize this change, I re-arranged the order of contexts
in the `.papr.yml` file in order to maintain inheritence between the
contexts and only pass in the extra args to the `centos/7/atomic`
context.