Vagrantfile: major updates to get current with test states
Originally, the `improved-sanity-test` required that the host under
test was at HEAD-1 of the stream we were testing.  Thus, the
Vagrantfile had additional scripts in place to get the box to the
proper state.

We've moved away from that methodology and now we can just `vagrant
up` a box and run our tests against that host.

This change updates the Vagrantfile to be current with the streams we
are most interested in and removes the additional provisioning steps
required to get the box to HEAD-1.

Closes #241