Add git to tools container
For the rpm-ostree compose test, we need the ability to checkout
a git repo with the necessary files to compose a tree.  Instead of
package layering git onto the host, we can use the tools container
with git installed to clone the repo and save it to a volume mount
shared between the host and container.