Install rpms for rpm-ostree install gpg test
Install createrepo_c and python on the aht-tools container so it we
can create a local repo and serve it through python's
SimpleHTTPServer.

Also added notes in the README to explain why the packages were added.
This might help in the future to determine if a package is still
needed in the tools container.