Add gpg package layering test
This commit adds a gpg test for package layering.  The test uses the
aht-tools container to create a local repo with the aht-dummy rpm and
serve it through SimpleHTTPServer.  The tools container was chosen
because it was an easy way to host the repo without requiring another
system or installing packages on the host.