i-s-t:  initial changes to support Fedora STI/STR
See the following for some details:

https://fedoraproject.org/wiki/CI/Tests
https://fedoraproject.org/wiki/CI/Standard_Test_Roles

These changes allow the `improved-sanity-test` playbook to be invoked
using the same pattern as described in the docs above.

The notable change is the value to `hosts:` which will try to find a
target under the `subjects` group (which is how the inventory file is
built via the STI).  But defaulting to `all` retains backward
compatibility with the original way of running the playbook.