reboot: modify to support Fedora STR
Additional details:

https://fedoraproject.org/wiki/CI/Tests
https://fedoraproject.org/wiki/CI/Standard_Test_Roles

When running any of the 'a-h-t' tests against an 'atomic' test subject
as defined by the Standard Test Interface, the harness will boot a VM
using `qemu` and fowards port 22 to a non-standard port (typically
I've seen 2222).  The side-effect of this is that the non-standard
port is always up and open, unless the VM has been torn down.  This
does *not* include when the host is rebooted.

This behavior causes our original methodology of waiting for the port
to go down and then come back up to be invalid.

I borrowed the idea of checking the `bootid` from the `rpm-ostree`
tests.  I also had to add in the ability to configure the `delay` on
the `wait_for` because occasionally the VM booted by the STI wouldn't
be ready within the default 30s.  Thankfully, this approach is
backwards compatible with the previous ways we have run the playbooks.

I'm still not 100% certain this is going to work all the time, but it
is a step in that direction.