roles: make 'set_is_atomic' a proper role
In 56d74f59, changes were made to adapt some of the roles for non
Atomic Hosts.  But this never amounted to anything, so I'm punting on
that effort and just going with an Atomic Host only policy.

These changes introduce `set_is_atomic` as a proper role and update
the other roles to use it as a dependency.

(Unfortunately, I re-indented the `redhat_subscription` role, so
everything looks changed there)