rpm_version: small modifications for Ansible 2.4+
When using this role under Ansible 2.4, I observed two errors:

1) the `include_tasks` portion was trying to include the
`version.yml` file from the current working directory, rather than the
tasks directory of the role

2) there was a deprecation warning about including variables at the
top of a task, however the warning never identified where this was
happening.  I was able to narrow it down to the `rpm_version` role and
the use of `include_tasks` without using `vars` to supply variables to
the task.

This commit fixes these issues like so:

- explicitly include the `version.yml` file from the tasks directory

- use the `vars` method of defining a variable for use in the
`include_task` statement