tree: remove use of atomic_install_version_check
This role was a hacky workaround while we waited for all the platforms
to get a current version of `atomic`.  By my measurement, even F26 has
v1.20 of `atomic`, so all of the platforms should support the normal
`atomic install --system` syntax.

(This also clears up an error I was getting related to using
`rpm_version` in the `dependencies` of the removed role)