hnof: make handler a tiny bit more resilient
Previously the handler was (trying) to perform some of the tasks on
the remote host.  This particular task is just setting the name of the
journal file, so we can do that locally and unprivileged.  This should
make it slightly more resilient to network problems (but only briefly
since we try to run `journalctl` on the host almost immediately
afterwards).

Additionally, I added the `quote` filter to the variables used by the
`shell` tasks for some addtional hardening.