reboot/main: increase default timeout
In RHBZ#1572944 it is speculated that /dev/random is taking a longer
amount of time to initialize with entropy in newer Linux kernels.
The observable result of this is that boot times are increased, to the
tune of 3-4m.  Perhaps I'm being overzealous with the 300s timeout,
but we don't have much to lose at this point.

h/t to @jlebon for the probable root cause in projectatomic/papr#97