rpm-ostree/livefs: use include_role in a few places
We were seeing instances where variable values passed to `import_role`
where not being respected each time it was called.  I thought that
each time an 'import' was done the changed variables would be
respected, but that doesn't appear to be the case.

Using `include_role` will allow us to dynamically load in variables
each time the role is used.  This is kind of troubling as we made a
large switch to using `include_role` in most places, but we've not run
into any problems like this yet.