move 'ostree admin status' to sanity test
In #381, a sanity check for `ostree admin status` was added to the
`rpm-ostree` test suite.  It is probably better suited to the 'sanity'
test suite so we are able to catch any regressions earlier in the
testing cycle.

I created a dumb role for the `ostree admin status` command since that
was easier than trying to wedge it into an existing role.