handler: also save journal of previous boot
This can be help when an error in a current boot happens because of
something that happened in the *previous* boot. There doesn't seem to be
a concise way to tell `journalctl` to list both the current and previous
boots but not the ones before that, so we just append to the tmpfile in
a second invocation.